+++ gemrb/core/Game.cpp
@@ -462,8 +462,8 @@ void Game::InitActorPos(Actor *actor) const
 	const char *area = start->QueryField(mode[playmode],"AREA");
 	const char *rot = start->QueryField(mode[playmode],"ROT");
 
-	actor->Pos.x = actor->Destination.x = (short) atoi( strta->QueryField( strta->GetRowIndex(xpos), ip ) );
-	actor->Pos.y = actor->Destination.y = (short) atoi( strta->QueryField( strta->GetRowIndex(ypos), ip ) );
+	actor->Pos.x = actor->Destination.x = atoi( strta->QueryField( strta->GetRowIndex(xpos), ip ));
+	actor->Pos.y = actor->Destination.y = atoi( strta->QueryField( strta->GetRowIndex(ypos), ip ));
 	actor->HomeLocation.x = actor->Pos.x;
 	actor->HomeLocation.y = actor->Pos.y;
 	actor->SetOrientation( atoi( strta->QueryField( strta->GetRowIndex(rot), ip) ), false );
